move more classes to common
