+++ failgood/common/src/failgood/Exceptions.kt
